Additions of swagger documentation to contact module -Zelealem Tesema
